+++ src/test/kotlin/blackjack/domain/DeckSpecs.kt
@@ -4,7 +4,7 @@ import blackjack.domain.Denomination.KING
 import blackjack.domain.Suit.SPADE
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.DescribeSpec
-import io.kotest.matchers.types.shouldBeSameInstanceAs
+import io.kotest.matchers.shouldBe
 
 class DeckSpecs : DescribeSpec({
 
@@ -12,10 +12,10 @@ class DeckSpecs : DescribeSpec({
         it("카드 한 장을 뽑을 수 있다") {
             val deck: Deck = CustomDeck(
                 listOf(
-                    Card.from(KING, SPADE),
+                    Card(KING, SPADE),
                 )
             )
-            deck.draw() shouldBeSameInstanceAs Card.from(KING, SPADE)
+            deck.draw() shouldBe Card(KING, SPADE)
         }
         context("덱에 카드가 남아있지 않은 경우") {
             it("카드를 뽑을 수 없다") {
