+++ src/test/kotlin/blackjack/domain/cards.kt
@@ -1,4 +1,4 @@
 package blackjack.domain
 
-fun cards(vararg card: Pair<Denomination, Suit>): List<Card> = card.toList().map { Card.from(it.first, it.second) }
-fun hand(vararg card: Pair<Denomination, Suit>): Hand = Hand(cards(*card))
+infix fun Denomination.to(suit: Suit) = Card(this, suit)
+fun hand(vararg cards: Card): Hand = Hand(cards.toList())
